parameritize everything so proto2 and proto3 have same coverage of tests.
