+++ test/python/memoize.py
@@ -4,6 +4,7 @@ import inspect
 from _pytest.fixtures import fail_fixturefunc
 
 
+cache = {}
 # Memoizes the value.
 # Works with generator or function.
 # has same restrictions on generators as pytest - can only yield once
@@ -11,24 +12,25 @@ from _pytest.fixtures import fail_fixturefunc
 # I think this is the behavior you want, start up, value, teardown happens once in expected scope.
 # The value is memoized.
 def memoize(func):
-    cache = {}
+    func_name = func.__name__
+    generator_key = func_name + '__generator__memoize__'
 
     def memoized_func(*args):
-        if 'value' not in cache:
+        if func_name not in cache:
             value = func(*args)
             if inspect.isgeneratorfunction(func):
-                cache['generator'] = value
+                cache[generator_key] = value
                 value = next(value)
-            cache['value'] = value
+            cache[func_name] = value
 
-        return cache['value']
+        return cache[func_name]
 
     def generator_after():
-        if 'generator' in cache:
+        if generator_key in cache:
             try:
-                next(cache['generator'])
+                next(cache[generator_key])
             except StopIteration:
-                cache.pop('generator')
+                cache.pop(generator_key)
             else:
                 fail_fixturefunc(func, "yield_fixture function has more than one 'yield'")
 
@@ -45,9 +47,12 @@ def memoize(func):
                     after()
             return %s
     '''
-    code = (code % (func.__name__, arg_string, arg_string, func.__name__))
+    code = (code % (func_name, arg_string, arg_string, func_name))
 
     generated = {}
     exec code.strip() in generated
 
     return generated["binder"](memoized_func, generator_after)
+
+def clear_memoized_cache():
+    cache.clear()
